+++ src/iterators/async_iterators/work_iter.rs
@@ -1,7 +1,7 @@
 #[allow(unused_imports)]
-use crate::{DetachedWorkIter, MRBIterator, WorkIter};
-use crate::AsyncDetachedWorkIter;
+use crate::{MRBIterator, WorkIter};
 use crate::iterators::async_iterators::async_macros::{futures_import, gen_common_futs, gen_fut, waker_registerer};
+use crate::iterators::async_iterators::AsyncIterator;
 use crate::iterators::util_macros::delegate;
 use crate::iterators::util_macros::muncher;
 use crate::ring_buffer::variants::ring_buffer_trait::{ConcurrentRB, MutRB};
@@ -18,20 +18,33 @@ pub struct AsyncWorkIter<'buf, B: MutRB> {
 }
 unsafe impl<'buf, B: ConcurrentRB + MutRB<Item = T>, T> Send for AsyncWorkIter<'buf, B> {}
 
-gen_common_futs! (&'a mut AsyncWorkIter<'buf, B>);
+impl<'buf, B: MutRB<Item = T>, T> AsyncIterator for AsyncWorkIter<'buf, B> {
+    type I = WorkIter<'buf, B>;
 
-impl<'buf, B: MutRB<Item = T>, T> AsyncWorkIter<'buf, B> {
-    pub fn from_sync(iter: WorkIter<'buf, B>) -> Self {
+    #[inline]
+    fn inner(&self) -> &Self::I {
+        &self.inner
+    }
+    #[inline]
+    fn inner_mut(&mut self) -> &mut Self::I {
+        &mut self.inner
+    }
+
+    fn into_sync(self) -> Self::I {
+        self.inner
+    }
+
+    fn from_sync(iter: Self::I) -> Self {
         Self {
             inner: iter,
             waker: None,
         }
     }
+}
 
-    pub fn into_sync(self) -> WorkIter<'buf, B> {
-        self.inner
-    }
+gen_common_futs! (&'a mut AsyncWorkIter<'buf, B>);
 
+impl<'buf, B: MutRB<Item = T>, T> AsyncWorkIter<'buf, B> {
     waker_registerer!();
     delegate!(WorkIter, pub fn is_prod_alive(&self) -> bool);
     delegate!(WorkIter, pub fn is_cons_alive(&self) -> bool);
@@ -41,13 +54,6 @@ impl<'buf, B: MutRB<Item = T>, T> AsyncWorkIter<'buf, B> {
     delegate!(WorkIter, pub unsafe fn advance(&(mut) self, count: usize));
     delegate!(WorkIter, pub fn available(&(mut) self) -> usize);
     delegate!(WorkIter, pub fn reset_index(&(mut) self));
-
-
-    /// Detaches the iterator yielding a [`DetachedWorkIter`].
-    #[inline]
-    pub fn detach(self) -> AsyncDetachedWorkIter<'buf, B> {
-        AsyncDetachedWorkIter::from_work(self)
-    }
 
     /// Returns a mutable references to the current value.
     ///