+++ src/iterators/mod.rs
@@ -18,12 +18,31 @@ pub(crate) mod util_macros {
     }
 
     macro_rules! delegate {
+        ($Inner: tt $(($inline: tt))?, $v: vis fn $fn_name: ident (&self $(, $arg: ident $(: $arg_t: ty)?)*)
+        $(-> $($ret_g: tt)*)?) => {
+            #[doc = concat!("Same as [`", stringify!($Inner), "::", stringify!($fn_name), "`].")]
+            $(#[$inline])?
+            $v fn $fn_name(&self $(, $arg $(: $arg_t)?)*) $(-> muncher!{ $($ret_g)* })? {
+                self.inner().$fn_name($($arg)*)
+            }
+        };
+        ($Inner: tt $(($inline: tt))?, $v: vis unsafe fn $fn_name: ident (&self $(, $arg: ident $(: $arg_t: ty)?)*)
+        $(-> $($ret_g: tt)*)?) => {
+            #[doc = concat!("Same as [`", stringify!($Inner), "::", stringify!($fn_name), "`].")]
+            /// # Safety
+            #[doc = concat!("Same as [`", stringify!($Inner), "::", stringify!($fn_name), "`].")]
+            $(#[$inline])?
+            $v unsafe fn $fn_name(&self $(, $arg $(: $arg_t)?)*) $(-> muncher!{ $($ret_g)* })? {
+                self.inner().$fn_name($($arg)*)
+            }
+        };
+        
         ($Inner: tt $(($inline: tt))?, $v: vis fn $fn_name: ident (& $(($m: tt))? self $(, $arg: ident $(: $arg_t: ty)?)*)
         $(-> $($ret_g: tt)*)?) => {
             #[doc = concat!("Same as [`", stringify!($Inner), "::", stringify!($fn_name), "`].")]
             $(#[$inline])?
             $v fn $fn_name(& $($m)? self $(, $arg $(: $arg_t)?)*) $(-> muncher!{ $($ret_g)* })? {
-                self.inner.$fn_name($($arg)*)
+                self.inner_mut().$fn_name($($arg)*)
             }
         };
         ($Inner: tt $(($inline: tt))?, $v: vis unsafe fn $fn_name: ident (& $(($m: tt))? self $(, $arg: ident $(: $arg_t: ty)?)*)
@@ -33,7 +52,7 @@ pub(crate) mod util_macros {
             #[doc = concat!("Same as [`", stringify!($Inner), "::", stringify!($fn_name), "`].")]
             $(#[$inline])?
             $v unsafe fn $fn_name(& $($m)? self $(, $arg $(: $arg_t)?)*) $(-> muncher!{ $($ret_g)* })? {
-                self.inner.$fn_name($($arg)*)
+                self.inner_mut().$fn_name($($arg)*)
             }
         };
     }